+++ include/openspace/interaction/keyframenavigator.h
