+++ include/openspace/interaction/orbitalnavigator.h
